pyproject.toml
[tool.poetry]
name = "achievements-bot"
version = "0.1.0"
description = ""
authors = ["Alexander Groza <a.groza.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "achievements_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.4"
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
